refactor: 변환 로직 분리
